Refactor: 폴더명 변경
